my-fullstack-agent/pyproject.toml
[project]
name = "deep-search"
version = "0.1.0"
description = ""
authors = [
    {name = "Amit Dutta", email = "amitdutta@google.com"},
    {name = "Deepak Moonat", email = "dmoonat@google.com"},
    {name = "Elia Secchi", email = "eliasecchi@google.com"},
    {name = "Erwin Huizenga", email = "erwinh@google.com"},
    {name = "Lavi Nigam", email = "lavinigam@google.com"},
]
dependencies = [
    "google-adk>=1.8.0",
    "python-docx>=1.1.0",
    "pdfplumber>=0.10.0",
]

requires-python = ">=3.10,<3.13"

[dependency-groups]
dev = [
    "pytest~=9.0.1",
    "pytest-asyncio~=1.3.0",
    "nest-asyncio>=1.6.0",
    "agent-starter-pack>=0.14.1",
]

[project.optional-dependencies]

jupyter = [
    "jupyter~=1.1.1",
]

lint = [
    "ruff>=0.4.6",
    "mypy~=1.19.0",
    "codespell~=2.4.1",
    "types-pyyaml~=6.0.12.20240917",
    "types-requests~=2.32.4.20250913",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["app", "frontend"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[tool.hatch.build.targets.wheel]
packages = ["app", "frontend", "deep_search", "resume_tailoring"]

# This configuration file is used by goo.gle/agent-starter-pack to power remote templating.
# It defines the template's properties, settings, and deployment options.
[tool.agent-starter-pack]
base_template = "adk_base"
description = "A production-ready fullstack research agent that uses Gemini to strategize, research, and synthesize comprehensive reports with human-in-the-loop collaboration"
example_question = "A report on the latest Google I/O event"

[tool.agent-starter-pack.settings]
deployment_targets = ["agent_engine", "cloud_run"]
interactive_command = "dev"
